Separate landing filter items into individual components
