+++ src/components/LandingFilter/KeywordInput/KeywordInput.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const KeywordInput = ({ value, onChange } = {}) => (
+  <input
+    name="keyword"
+    value={value}
+    onChange={onChange}
+    placeholder="Type Keywords"
+  />
+);
+
+export default KeywordInput;
