+++ src/components/LandingFilter/PriceSlider/PriceSlider.styles.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const CategoryHeader = styled.span`
+  display: block;
+  width: 100%;
+  text-align: end;
+`;
+
+export default CategoryHeader;
